refactoring content components
