+++ src/components/Attestations/AttestationFormContent.jsx
@@ -1,17 +1,17 @@
 /* eslint-disable react/prop-types */
 import React from 'react'
-import Attest from './Attest'
-import ReadAttestation from './ReadAttestation'
-import AboutAttestations from './AboutAttestations'
+import NewAttestation from '../NewAttestation'
+import ReadAttestation from '../ReadAttesatation'
+import About from '../About'
 
 const AttestationFormContent = (props) => {
   switch (props.active) {
     case 0:
-      return <Attest />
+      return <NewAttestation />
     case 1:
       return <ReadAttestation />
     case 2:
-      return <AboutAttestations />
+      return <About />
     default:
       return <div>How&apos;d you get here???</div>
   }
